backend/requirements.txt
fastapi-clerk-auth
python-jose[cryptography]
sqlmodel
openai
python-dotenv